message length double precision added and incremented

+++ pkg/timeseries/consumer.go
@@ -91,7 +91,7 @@ func (h *Consumer) ConsumeClaim(s sarama.ConsumerGroupSession, claim sarama.Cons
 				Property:  property,
 				Timestamp: msg.Timestamp,
 				Value:     datapointValue,
-				Length:    len(message.Value),
+				Length:    float32(len(message.Value)),
 			})
 
 		}
